it works! renders a rectangle

+++ src/com/bubble/hearthstone/module/render/opengl/IGraphics.java
@@ -1,5 +1,5 @@
 package com.bubble.hearthstone.module.render.opengl;
 
 public interface IGraphics {
-    void drawRect(int x ,int y, int w, int h);
+    void drawRect(float x, float y, float w, float h);
 }
